package reorganization: moved modules to utilities package
