+++ nmt_chainer/models/attention.py
@@ -15,7 +15,7 @@ import chainer.functions as F
 import chainer.links as L
 import math, random
 
-from nmt_chainer.utils import ortho_init, compute_lexicon_matrix, minibatch_sampling
+from nmt_chainer.utilities.utils import ortho_init, compute_lexicon_matrix, minibatch_sampling
 
 import logging
 logging.basicConfig()
