+++ nmt_chainer/models/decoder_cells.py
@@ -15,9 +15,9 @@ import chainer.links as L
 import math, random
 
 import rnn_cells
-from nmt_chainer.utils import ortho_init, minibatch_sampling
+from nmt_chainer.utilities.utils import ortho_init, minibatch_sampling
 
-from nmt_chainer.constant_batch_mul import batch_matmul_constant, matmul_constant
+from nmt_chainer.utilities.constant_batch_mul import batch_matmul_constant, matmul_constant
 
 from attention import AttentionModule
