+++ nmt_chainer/scripts/extract_dict.py
