+++ nmt_chainer/scripts/lex_dict.py
