+++ nmt_chainer/utilities/bleu_computer.py
