+++ nmt_chainer/utilities/constant_batch_mul.py
