+++ nmt_chainer/utilities/faster_gru.py
