+++ nmt_chainer/utilities/utils.py
