+++ nmt_chainer/utilities/versioning_tools.py
