+++ nmt_chainer/utilities/visualisation.py
